+++ src/main/java/com/example/products/repository/CompanyRepository.java
@@ -1,11 +1,11 @@
 package com.example.products.repository;
 
-import com.example.products.data.Company;
+import com.example.products.entity.Company;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CompanyRepository extends JpaRepository<Company, Integer> {
+public interface CompanyRepository extends JpaRepository<Company, Long> {
 
-Company getCompanyById(int id);
+Company getCompanyById(long id);
 }
