+++ src/main/java/com/example/products/repository/ProductRepository.java
@@ -1,17 +1,16 @@
 package com.example.products.repository;
 
-import com.example.products.data.Product;
+import com.example.products.entity.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import javax.print.attribute.standard.PDLOverrideSupported;
 import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface ProductRepository extends JpaRepository<Product, Integer> {
+public interface ProductRepository extends JpaRepository<Product, Long> {
 
-    List<Product>  getProductsByCompany_Id(int id);
+    List<Product>  getProductsByCompany_Id(long id);
 
    Optional<Product>  getProductByName(String name);
