+++ src/main/java/com/example/products/repository/WishListRepo.java
@@ -1,7 +1,8 @@
 package com.example.products.repository;
 
-import com.example.products.data.WishList;
+import com.example.products.entity.WishList;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface WishListRepo extends JpaRepository<WishList, Integer> {
+public interface WishListRepo extends JpaRepository<WishList, Long> {
+
 }
